facto oauth uni_test code
